Refactor Falcon API and tasks into their own respective modules
